+++ src/main/java/com/example/bookbackend/common/util/ResponseManager.java
@@ -1,7 +1,7 @@
-package com.example.bookbackend.common.response;
+package com.example.bookbackend.common.util;
 
 import com.example.bookbackend.auth.web.dto.Tokens;
-import com.example.bookbackend.common.exception.CommonResponse;
+import com.example.bookbackend.common.response.CommonResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
