+++ src/main/java/com/example/bookbackend/member/application/MemberService.java
@@ -9,7 +9,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import static com.example.bookbackend.common.util.ApiCode.*;
+import static com.example.bookbackend.common.response.ApiCode.*;
 
 @RequiredArgsConstructor
 @Transactional
